DukaanAPI/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
idna==3.4
pycodestyle==2.11.0
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
